Added ability to specify individual series 'type'
Added TYPE to HighchartsSeries and enum for valid types. Series
instantiated without TYPE specified will retain the type from the chart,
Those instantiated with new constructor will display that series with
the indicated type. Added an 'areaspline' to the sample controller to
demonstrate. Not for public consumption in this format.